Travis CI: add .travis.yml to enable Travis CI
This commit adds a configuration file for Travis CI (travis-ci.org). Travis CI
is a continuous integration platform which will allow us to build qfs on every
commit. While we have internal continuous integration within Quantcast, it's
also important to use external continuous integration as well. It means that
pull requests submitted by external parties will be built and tested before we
review them. It also means that we can ensure that all code we push upstream
will be built and validated on external infrastructure.

The qfs tests are currently run through qfstest.sh. This fails on travis ci with
an error:

    ../../src/test-scripts/qfstest.sh: 533: kill: No such process

We have a plan to move the qfs tests to use a proper testing framework. Until
that can be done or we can get the current unit tests to work on Travis CI, unit
tests are disabled.

http://reviews.sfoffice.qc/r/41133/