build: prevent relinking of all binaries on every build
The qfs build system generates a Version.cc file in src/cc/common every time the
project is built. This file includes various configuration settings like your
git configuration, the tag being built from, etc. It also includes whether the
working directory was dirty when building and other useful information to help
debug how a binary was made. However, it always overwrites the Version.cc file
with the new one. This means that the common library has to be rebuilt. So does
every other library and binary which depends on the common library. As the name
suggests, quite a few libraries and binaries depend on the common library. The
result is almost all of the project's libraries and binaries being relinked.
It's an unnecessary build step that wastes time. This change fixes that so that
we don't unnecessarily relink all the libraries and binaries in the project.

http://reviews.sfoffice.qc/r/41362/