wiki/Home.md: add a note on how to update the qfs wiki with our new process
The community is encouraged to update the QFS wiki as they wish. However, the
best that GitHub offers for this is to allow public editing. Public editing is
great, but it doesn't have a review process. Furthermore, GitHub doesn't support
pull requests for wikis. Instead, the wiki documents are now mirrored in the
`wiki` directory of the QFS source code. This allows for a few things:

- Wiki documentation is distributed along with the qfs source code
- The community can use pull requests to submit changes to the wiki
- The community can request updates to the wiki in the same pull request as code
  changes are made, keeping code and documentation in sync