travis: disable email notifications to qfs-dev@quantcast.com
By default, a build email is sent to the committer and the author, but only if
they have access to the repository the commit was pushed to. This prevents forks
active on Travis CI from notifying the upstream repository’s owners when they’re
pushing any upstream changes to their fork. It also prevents build notifications
from going to folks not registered on Travis CI.

As we get more and more people forking QFS, we don't want to get build failure
emails from all of their forks. Let's remove this since it didn't work the way I
wanted it to, unfortunately.