tests: add google test to the qfs source
This change adds google test to the qfs source code. It updates the cmake files
so that all tests are built and run under the make gtest command. This testing
framework comes with proper assertions, fixtures, environments, etc. It is a
much more resourceful framework than what we currently have with qftest.sh, e.g.
tests are run in isolation using a syntax and API which isn't scary to
developers unfamiliar with larger parts of the codebase. This change starts with
adding the framework to the build system. Over the next couple of commits, we
will work on migrating the qfstest.sh tests to run under the google test
framework.